# swagger/cla.yaml
...
swagger: '2.0'
info:
  title: EasyCLA API
  version: '1.0.0'
basePath: /v4
securityDefinitions:
  lf-auth:
    type: apiKey
    name: X-ACL
    in: header
    description: Requires X-ACL headers and bearer token
security:
  - lf-auth: []
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

#securityDefinitions:
#  OauthSecurity:
#    type: oauth2
#    flow: accessCode
#    authorizationUrl: 'https://accounts.google.com/o/oauth2/v2/auth'
#    tokenUrl: 'https://www.googleapis.com/oauth2/v4/token'
#    scopes:
#      admin: LF Admin scope
#      project: Project Manager scope
#      company: CLA Manager scope
#      contrib: Contributor scope

#security:
#  - OauthSecurity: []
...
paths:
...
  /repositories:
    post:
      summary: create repositories
      description: create repositories
      operationId: createRepositories
      parameters:
        - $ref: '#/parameters/x-acl'
        - $ref: '#/parameters/x-username'
        - $ref: '#/parameters/x-email'
        - in: body
          description: repositories to create
          name: repositories_input
          schema:
            $ref: '#/definitions/create_repositories_input'
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/repository_list'
        '400':
          $ref: '#/responses/invalid-request'
        '401':
          $ref: '#/responses/unauthorized'
        '403':
          $ref: '#/responses/forbidden'
      tags:
        - repositories
    get:
      summary: get repositories list
      description: return list of repositores
      operationId: listRepositories
      parameters:
        - $ref: '#/parameters/x-acl'
        - $ref: '#/parameters/x-username'
        - $ref: '#/parameters/x-email'
        - $ref: '#/parameters/cla_group_id'
        - $ref: '#/parameters/list_repositories_project_id'
        - $ref: '#/parameters/repository_name'
        - $ref: '#/parameters/repository_type'
        - $ref: '#/parameters/repository_organization'
        - $ref: '#/parameters/repositories_order_by'
        - $ref: '#/parameters/sort_order'
        - $ref: '#/parameters/offset'
        - $ref: '#/parameters/page_size'
      produces:
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/repository_list'
        '400':
          $ref: '#/responses/invalid-request'
        '401':
          $ref: '#/responses/unauthorized'
        '403':
          $ref: '#/responses/forbidden'
      tags:
        - repositories
    delete:
      summary: delete repositories for given cla_group and project
      description: delete repositories
      operationId: deleteRepositories
      parameters:
        - $ref: '#/parameters/x-acl'
        - $ref: '#/parameters/x-username'
        - $ref: '#/parameters/x-email'
        - in: body
          description: repositories to create
          name: repositories_input
          schema:
            $ref: '#/definitions/delete_repositories_input'
      produces:
        - application/json
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/responses/invalid-request'
        '401':
          $ref: '#/responses/unauthorized'
        '403':
          $ref: '#/responses/forbidden'
        '404':
          $ref: '#/responses/not-found'
      tags:
        - repositories
definitions:
...
  error-response:
    type: object
    title: Error Response
    description: Standard error format
    properties:
      Code:
        type: string
        description: The code for the error response
        example: "403"
        pattern: '^([1-9][\d]+){1,}$'
      Message:
        type: string
        description: The message for the error response
        example: "Not authorized"
        pattern: '^([\w\d\s\-\,\./]+){2,}$'
...
  repository_list:
    title: repository_list
    type: object
    properties:
      repositories:
        type: array
        items:
          $ref: '#/definitions/repository'
  repository:
    title: repository
    type: object
    properties:
      id:
        type: string
      repository_type:
        type: string
        enum:
          - github
          - gerrit
          - gitlab
      name:
        type: string
      organization_name:
        type: string
      external_id:
        type: string
      url:
        type: string
      enabled:
        type: boolean
      project_id:
        type: string
      cla_group_id:
        type: string
      created_at:
        type: integer
  create_repositories_input:
    title: create_repositories_input
    type: object
    required:
      - project_id
      - cla_group_id
    properties:
      project_id:
        type: string
      cla_group_id:
        type: string
      repositories:
        type: array
        items:
          type: object
          minItems: 1
          required:
            - name
            - organization_name
            - url
            - external_id
          properties:
            repository_type:
              type: string
              enum:
                - github
                - gerrit
                - gitlab
            name:
              type: string
            organization_name:
              type: string
            url:
              type: string
            external_id:
              type: string
            enabled:
              type: boolean
              default: true
  delete_repositories_input:
    title: delete_repositories_input
    type: object
    required:
      - project_id
      - cla_group_id
    properties:
      project_id:
        type: string
      cla_group_id:
        type: string
      repository_ids:
        type: array
        minItems: 1
        items:
          type: string
responses:
  unauthorized:
    description: Unauthorized
    schema:
      $ref: '#/definitions/error-response'
  invalid-request:
    description: Invalid request
    schema:
      $ref: '#/definitions/error-response'
  forbidden:
    description: Insufficient privilege to execute action.
    schema:
      $ref: '#/definitions/error-response'
  not-found:
    description: Not found
    schema:
      $ref: '#/definitions/error-response'
  #conflict:
  #  description: The request could not be completed due to a conflict with the current state of the target resource.
  #  schema:
  #    $ref: '#/definitions/error-response'
parameters:
  page_size:
    name: page_size
    description: The maximum number of results per page, value must be a positive integer value
    in: query
    type: integer
    default: 100
  offset:
    name: offset
    description: The page offset for fetching subsequent pages of results, value must be a non-negative integer value
    in: query
    type: integer
    default: 0
...
  sort_order:
    name: sort_order
    in: query
    description: Sort order
    type: string
    enum: [asc, desc]
...
  list_repositories_project_id:
    name: project_id
    description: unique id of the project
    in: query
    type: string
    required: true
...
  cla_group_id:
    name: cla_group_id
    description: unique id of the cla_group
    in: query
    type: string
  repository_name:
    name: repository_name
    description: name of the repository
    in: query
    type: string
  repository_type:
    name: repository_type
    description: type of the repository
    in: query
    type: string
  repository_organization:
    name: repository_organization
    description: organization of the repository
    in: query
    type: string
  repositories_order_by:
    name: order_by
    description: The field to order by, default sort order is ascending
    in: query
    type: string
    enum: [name,organization_name]
...
  x-acl:
    name: X-ACL
    description: The access control list header value encoded as base64 - assigned by the API Gateway based on user/request permissions
    in: header
    type: string
    required: true
  x-username:
    name: X-USERNAME
    description: The username/LFID value - assigned/set by the API Gateway based on login credentials
    in: header
    type: string
  x-email:
    name: X-EMAIL
    description: The email value - assigned/set by the API Gateway based on login credentials
    in: header
    type: string